requirements.txt
python-osc
pyautogui
pillow
numpy
pywin32